Retabbed the file using VIM's automation features.
Did this because applying code changes to a standard that uses actual
tabs and not four spaces was a bit hard on my editor. Hoping the
original author doesn't mind but I did not want to change my environment
just to adapt this script and I personally find that the four spaces
ideal is more robust.